Let the CSS engine do the hard work
So far we did calculate the positions of the "media" controls manually,
but this requires to actually use images whose size can be detected via
`getimagesize()`.  Fortunately, nowadays we can use a flexbox which
lays out the controls as we desire without any manual positioning.  The
controls are still usable in old browsers (e.g. IE ≤ 9), albeit they
are placed in the upper left corner of the image.